Create separate bpmn file
